+++ CHANGELOG.md
@@ -23,6 +23,8 @@ need to address.
 - ASCII logo.
 - Version line after ASCII logo.
 - getopts section for main command.
+- Help option.
+- Error message when an invalid option is received.
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html